+++ tests/Silex/Tests/Application/SecurityTraitTest.php
@@ -13,7 +13,6 @@ namespace Silex\Tests\Application;
 
 use Silex\Provider\SecurityServiceProvider;
 use Symfony\Component\Security\Core\User\User;
-use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -23,47 +22,6 @@ use Symfony\Component\HttpFoundation\Request;
  */
 class SecurityTraitTest extends \PHPUnit_Framework_TestCase
 {
-    public function testUser()
-    {
-        $request = Request::create('/');
-
-        $app = $this->createApplication(array(
-            'fabien' => array('ROLE_ADMIN', '5FZ2Z8QIkA7UTZ4BYkoC+GsReLf569mSKDsfods6LYQ8t+a8EW9oaircfMpmaLbPBh4FOBiiFyLfuZmTSUwzZg=='),
-        ));
-        $app->get('/', function () { return 'foo'; });
-        $app->handle($request);
-        $this->assertNull($app->user());
-
-        $request->headers->set('PHP_AUTH_USER', 'fabien');
-        $request->headers->set('PHP_AUTH_PW', 'foo');
-        $app->handle($request);
-        $this->assertInstanceOf('Symfony\Component\Security\Core\User\UserInterface', $app->user());
-        $this->assertEquals('fabien', $app->user()->getUsername());
-    }
-
-    public function testUserWithNoToken()
-    {
-        $request = Request::create('/');
-
-        $app = $this->createApplication();
-        $app->get('/', function () { return 'foo'; });
-        $app->handle($request);
-        $this->assertNull($app->user());
-    }
-
-    public function testUserWithInvalidUser()
-    {
-        $request = Request::create('/');
-
-        $app = $this->createApplication();
-        $app->boot();
-        $app['security.token_storage']->setToken(new UsernamePasswordToken('foo', 'foo', 'foo'));
-
-        $app->get('/', function () { return 'foo'; });
-        $app->handle($request);
-        $this->assertNull($app->user());
-    }
-
     public function testEncodePassword()
     {
         $app = $this->createApplication(array(